add toJSON instances for Resolved statements
This will permit the implementation (externally) of a service endpoint
"parseAndResolve" which returns a query's resolved AST.

There were a few toJSON instances that had to be implemented. In the
future, we will catch missing toJSON-instances at compile time, thanks
to the added explicit type examples on Vertica/Hive/Presto statements.